split UpdateProfileModal
Split UpdateProfileModal content into two separate components, one for
profile image and one for personal info
