+++ src/components/userProfile/userProfileMainContent/user/updateProfileModal/profileImageSection/profileImageSection.module.scss
@@ -1,0 +1,66 @@
+.profileImage {
+  display: block;
+  width: 100%;
+  background-color: #bbb;
+  -webkit-box-shadow: 0px 0px 5px 1px #ddd;
+  box-shadow: 0px 0px 5px 1px #ddd;
+}
+
+.profileImageButtons {
+  margin: 1.2em 0 1.7em;
+}
+
+.button {
+  width: 100%;
+  font-size: 15px;
+  padding: 0.5em 0;
+  margin: 0.8em 0;
+}
+
+@media (min-width: 550px) and (orientation: portrait) {
+  .profileImage {
+    width: 70%;
+    margin: 0 auto;
+  }
+  .profileImageButtons {
+    display: flex;
+    justify-content: space-between;
+  }
+
+  .button {
+    flex-basis: 47%;
+    font-size: 18px;
+  }
+}
+
+@media (min-width: 768px) and (orientation: portrait) {
+  .profileImage {
+    width: 65%;
+  }
+}
+
+@media (min-width: 600px) and (orientation: landscape) {
+  .profileImageSec {
+    flex-basis: 38%;
+  }
+
+  .profileImageButtons {
+    margin: 0.9em 0 0em;
+  }
+}
+
+@media (min-width: 800px) and (orientation: landscape) {
+  .profileImageSec {
+    flex-basis: 35%;
+  }
+}
+
+@media (min-width: 1000px) and (orientation: landscape) {
+  .profileImageSec {
+    flex-basis: 35%;
+  }
+
+  .button {
+    font-size: 18px;
+  }
+}
