+++ src/components/userProfile/userProfileMainContent/user/updateProfileModal/updateProfileModal.module.scss
@@ -12,77 +12,15 @@
   overflow-y: auto;
 }
 
-.profileImage {
-  display: block;
-  width: 100%;
-  background-color: #bbb;
-  -webkit-box-shadow: 0px 0px 5px 1px #ddd;
-  box-shadow: 0px 0px 5px 1px #ddd;
-}
-
-.profileImageButtons {
-  margin: 1.2em 0 1.7em;
-}
-
-.header {
-  font-size: 25px;
-  font-weight: bold;
-}
-
-.inputsContainer {
-  margin: 0.9em 0 1.4em;
-}
-
-.button {
-  width: 100%;
-  font-size: 15px;
-  padding: 0.5em 0;
-  margin: 0.8em 0;
-}
-
-.input {
-  margin: 0.3em 0;
-  width: 100%;
-}
-
 @media (min-width: 550px) and (orientation: portrait) {
   .updateProfileModalComponent {
     max-width: 480px;
-  }
-  .profileImage {
-    width: 70%;
-    margin: 0 auto;
-  }
-  .profileImageButtons {
-    display: flex;
-    justify-content: space-between;
-  }
-
-  .profileImageButtons .button {
-    flex-basis: 47%;
-  }
-
-  .header {
-    font-size: 30px;
-  }
-
-  .button {
-    font-size: 18px;
-  }
-
-  form .button {
-    width: 50%;
-    margin: 1em auto;
   }
 }
 
 @media (min-width: 768px) and (orientation: portrait) {
   .updateProfileModalComponent {
     max-width: 580px;
-  }
-
-  .profileImage {
-    width: 65%;
   }
 }
 
@@ -95,68 +33,17 @@
     display: flex;
     justify-content: space-between;
   }
-  .profileImageSec {
-    flex-basis: 38%;
-  }
-
-  .profileInfoSec {
-    flex-basis: 50%;
-  }
-
-  .profileImageButtons {
-    margin: 0.9em 0 0em;
-  }
-
-  .header {
-    display: none;
-  }
-
-  .inputsContainer {
-    margin: 0 0 1em;
-  }
-
-  form .button {
-    width: 60%;
-    margin-left: auto;
-  }
 }
 
 @media (min-width: 800px) and (orientation: landscape) {
   .updateProfileModalComponent {
     width: 80%;
   }
-  .profileImageSec {
-    flex-basis: 35%;
-  }
-
-  .profileInfoSec {
-    flex-basis: 52%;
-  }
-
-  form .button {
-    width: 50%;
-  }
 }
 
 @media (min-width: 1000px) and (orientation: landscape) {
   .updateProfileModalComponent {
     max-width: 800px;
-  }
-
-  .profileImageSec {
-    flex-basis: 35%;
-  }
-
-  .profileInfoSec {
-    flex-basis: 50%;
-  }
-
-  .button {
-    font-size: 18px;
-  }
-
-  form .button {
-    width: 45%;
   }
 }
